# Perf Improvements to Custom Commands' Execution Path