--- a/libs/server/Custom/CustomCommandManager.cs
+++ b/libs/server/Custom/CustomCommandManager.cs
@@ -3,6 +3,7 @@
 
 using System;
 using System.Collections.Concurrent;
+using System.Collections.Generic;
 using System.Diagnostics;
 using Microsoft.Extensions.Logging;
 
@@ -13,26 +14,35 @@ namespace Garnet.server
     /// </summary>
     public class CustomCommandManager
     {
+        // Maximum number of custom raw string commands
         private static readonly int MaxCustomRawStringCommands = 256;
 
+        // Initial size of expandable maps
+        private static readonly int MinMapSize = 8;
+
+        // ID ranges for custom raw string commands and custom object types
         private static readonly int CustomRawStringCommandMinId = (ushort)RespCommand.INVALID - MaxCustomRawStringCommands;
         private static readonly int CustomRawStringCommandMaxId = (ushort)RespCommand.INVALID - 1;
         private static readonly int CustomObjectTypeMinId = (byte)GarnetObjectTypeExtensions.LastObjectType + 1;
         private static readonly int CustomObjectTypeMaxId = (byte)GarnetObjectTypeExtensions.FirstSpecialObjectType - 1;
 
+        // Maps holding different types of custom commands
         private ConcurrentExpandableMap<CustomRawStringCommand> rawStringCommandMap;
         private ConcurrentExpandableMap<CustomObjectCommandWrapper> objectCommandMap;
         private ConcurrentExpandableMap<CustomTransaction> transactionProcMap;
         private ConcurrentExpandableMap<CustomProcedureWrapper> customProcedureMap;
 
+        // Map holding all registered modules by module name
         private readonly ConcurrentDictionary<string, ModuleLoadContext> modules;
 
-        internal static readonly int MinMapSize = 8;
-        internal static readonly byte CustomTypeIdStartOffset = (byte)CustomObjectTypeMinId;
+        // Maps holding custom command info and docs by command name
+        private readonly ConcurrentDictionary<string, RespCommandsInfo> customCommandsInfo = new(StringComparer.OrdinalIgnoreCase);
+        private readonly ConcurrentDictionary<string, RespCommandDocs> customCommandsDocs = new(StringComparer.OrdinalIgnoreCase);
 
-        internal int CustomCommandsInfoCount => customCommandsInfo.Count;
-        internal readonly ConcurrentDictionary<string, RespCommandsInfo> customCommandsInfo = new(StringComparer.OrdinalIgnoreCase);
-        internal readonly ConcurrentDictionary<string, RespCommandDocs> customCommandsDocs = new(StringComparer.OrdinalIgnoreCase);
+        /// <summary>
+        /// Offset from which custom type IDs start in the GarnetObjectType enum
+        /// </summary>
+        internal static readonly byte CustomTypeIdStartOffset = (byte)CustomObjectTypeMinId;
 
         /// <summary>
         /// Create new custom command manager
@@ -53,22 +63,45 @@ namespace Garnet.server
             modules = new();
         }
 
+        /// <summary>
+        /// Register custom raw-string command
+        /// </summary>
+        /// <param name="name">Name of command</param>
+        /// <param name="type">Command type</param>
+        /// <param name="customFunctions">Custom raw-string functions</param>
+        /// <param name="commandInfo">Command info of command</param>
+        /// <param name="commandDocs">Command docs of command</param>
+        /// <param name="expirationTicks">Command expiration in ticks</param>
+        /// <returns>Command ID</returns>
+        /// <exception cref="Exception"></exception>
         internal int Register(string name, CommandType type, CustomRawStringFunctions customFunctions,
             RespCommandsInfo commandInfo, RespCommandDocs commandDocs, long expirationTicks)
         {
             if (!rawStringCommandMap.TryGetNextId(out var cmdId))
                 throw new Exception("Out of registration space");
             Debug.Assert(cmdId <= ushort.MaxValue);
+
             var extId = cmdId - CustomRawStringCommandMinId;
             var arity = commandInfo?.Arity ?? 0;
             var newCmd = new CustomRawStringCommand(name, (ushort)extId, type, arity, customFunctions, expirationTicks);
             var setSuccessful = rawStringCommandMap.TrySetValue(cmdId, ref newCmd);
             Debug.Assert(setSuccessful);
+
             if (commandInfo != null) customCommandsInfo.AddOrUpdate(name, commandInfo, (_, _) => commandInfo);
             if (commandDocs != null) customCommandsDocs.AddOrUpdate(name, commandDocs, (_, _) => commandDocs);
+            
             return extId;
         }
 
+        /// <summary>
+        /// Register custom transaction
+        /// </summary>
+        /// <param name="name">Name of transaction</param>
+        /// <param name="proc">Delegate for creating transaction</param>
+        /// <param name="commandInfo">Command info of transaction</param>
+        /// <param name="commandDocs">Command docs of transaction</param>
+        /// <returns>Command ID</returns>
+        /// <exception cref="Exception"></exception>
         internal int Register(string name, Func<CustomTransactionProcedure> proc, RespCommandsInfo commandInfo = null,
             RespCommandDocs commandDocs = null)
         {
@@ -80,11 +113,19 @@ namespace Garnet.server
             var newCmd = new CustomTransaction(name, (byte)cmdId, arity, proc);
             var setSuccessful = transactionProcMap.TrySetValue(cmdId, ref newCmd);
             Debug.Assert(setSuccessful);
+
             if (commandInfo != null) customCommandsInfo.AddOrUpdate(name, commandInfo, (_, _) => commandInfo);
             if (commandDocs != null) customCommandsDocs.AddOrUpdate(name, commandDocs, (_, _) => commandDocs);
+            
             return cmdId;
         }
 
+        /// <summary>
+        /// Register custom object type
+        /// </summary>
+        /// <param name="factory">Factory for creating custom object</param>
+        /// <returns>Type ID</returns>
+        /// <exception cref="Exception"></exception>
         internal int RegisterType(CustomObjectFactory factory)
         {
             if (objectCommandMap.TryGetFirstId(c => c.factory == factory, out var dupRegistrationId))
@@ -94,6 +135,17 @@ namespace Garnet.server
             return typeId - CustomObjectTypeMinId;
         }
 
+        /// <summary>
+        /// Register custom object command
+        /// </summary>
+        /// <param name="name">Name of command</param>
+        /// <param name="commandType">Command type</param>
+        /// <param name="factory">Factory for creating custom object</param>
+        /// <param name="commandInfo">Command info for custom command</param>
+        /// <param name="commandDocs">Command docs for custom command</param>
+        /// <param name="customObjectFunctions">Custom object functions</param>
+        /// <returns>Object type ID and sub-command ID</returns>
+        /// <exception cref="Exception"></exception>
         internal (int objectTypeId, int subCommand) Register(string name, CommandType commandType, CustomObjectFactory factory, RespCommandsInfo commandInfo, RespCommandDocs commandDocs, CustomObjectFunctions customObjectFunctions = null)
         {
             if (!objectCommandMap.TryGetFirstId(c => c.factory == factory, out var typeId))
@@ -106,8 +158,8 @@ namespace Garnet.server
             objectCommandMap.TryGetValue(typeId, out var wrapper);
             if (!wrapper.commandMap.TryGetNextId(out var scId))
                 throw new Exception("Out of registration space");
-
             Debug.Assert(scId <= byte.MaxValue);
+
             var arity = commandInfo?.Arity ?? 0;
             var newSubCmd = new CustomObjectCommand(name, (byte)extId, (byte)scId, commandType, arity, wrapper.factory,
                 customObjectFunctions);
@@ -121,13 +173,13 @@ namespace Garnet.server
         }
 
         /// <summary>
-        /// Register custom command
+        /// Register custom procedure
         /// </summary>
-        /// <param name="name"></param>
-        /// <param name="customProcedure"></param>
-        /// <param name="commandInfo"></param>
-        /// <param name="commandDocs"></param>
-        /// <returns></returns>
+        /// <param name="name">Name of custom procedure</param>
+        /// <param name="customProcedure">Delegate for creating custom procedure</param>
+        /// <param name="commandInfo">Command info of custom procedure</param>
+        /// <param name="commandDocs">Command docs of custom procedure</param>
+        /// <returns>Command ID</returns>
         /// <exception cref="Exception"></exception>
         internal int Register(string name, Func<CustomProcedure> customProcedure, RespCommandsInfo commandInfo = null, RespCommandDocs commandDocs = null)
         {
@@ -182,53 +234,151 @@ namespace Garnet.server
             return true;
         }
 
+        /// <summary>
+        /// Try to add a module to the modules map (should only be called by ModuleRegistrar)
+        /// </summary>
+        /// <param name="moduleLoadContext">Module load context</param>
+        /// <returns>True if module was added successfully (i.e. its name doesn't already exist)</returns>
         internal bool TryAddModule(ModuleLoadContext moduleLoadContext)
             => modules.TryAdd(moduleLoadContext.Name, moduleLoadContext);
 
-        internal bool TryGetCustomProcedure(int id, out CustomProcedureWrapper value)
-            => customProcedureMap.TryGetValue(id, out value);
+        /// <summary>
+        /// Try to retrieve a custom procedure wrapper by procedure ID (should only be called by CustomCommandManagerSession)
+        /// </summary>
+        /// <param name="id">Procedure ID</param>
+        /// <param name="value">Retrieved custom procedure</param>
+        /// <returns>True if custom procedure found</returns>
+        internal bool TryGetCustomProcedure(int id, out CustomProcedureWrapper value)  
+            => customProcedureMap.eMapUnsafe.TryGetValue(id, out value);
 
+        /// <summary>
+        /// Try to retrieve a custom transaction by transaction ID (should only be called by CustomCommandManagerSession)
+        /// </summary>
+        /// <param name="id">Transaction ID</param>
+        /// <param name="value">Retrieved custom transaction</param>
+        /// <returns>True if custom transaction found</returns>
         internal bool TryGetCustomTransactionProcedure(int id, out CustomTransaction value)
-            => transactionProcMap.TryGetValue(id, out value);
+            => transactionProcMap.eMapUnsafe.TryGetValue(id, out value);
 
+        /// <summary>
+        /// Try to retrieve a custom raw string command by command ID
+        /// </summary>
+        /// <param name="id">Command ID</param>
+        /// <param name="value">Retrieved command</param>
+        /// <returns>True if command found</returns>
         internal bool TryGetCustomCommand(int id, out CustomRawStringCommand value)
-            => rawStringCommandMap.TryGetValue(id, out value);
+            => rawStringCommandMap.eMapUnsafe.TryGetValue(id, out value);
 
+        /// <summary>
+        /// Try to retrieve a custom object command by command ID
+        /// </summary>
+        /// <param name="id">Command ID</param>
+        /// <param name="value">Retrieved command</param>
+        /// <returns>True if command found</returns>
         internal bool TryGetCustomObjectCommand(int id, out CustomObjectCommandWrapper value)
-            => objectCommandMap.TryGetValue(id, out value);
+            => objectCommandMap.eMapUnsafe.TryGetValue(id, out value);
 
+        /// <summary>
+        /// Try to retrieve a custom object sub-command command by command ID and sub-command ID
+        /// </summary>
+        /// <param name="id">Command ID</param>
+        /// <param name="subId">Sub-command ID</param>
+        /// <param name="value">Retrieved command</param>
+        /// <returns>True if command found</returns>
         internal bool TryGetCustomObjectSubCommand(int id, int subId, out CustomObjectCommand value)
         {
-            value = default;
-            return objectCommandMap.TryGetValue(id, out var wrapper) &&
-                   wrapper.commandMap.TryGetValue(subId, out value);
+            value = null;
+            return objectCommandMap.eMapUnsafe.TryGetValue(id, out var wrapper) &&
+                   wrapper.commandMap.eMapUnsafe.TryGetValue(subId, out value);
         }
 
+        /// <summary>
+        /// Get a custom raw-string command by name
+        /// </summary>
+        /// <param name="command">The command name to match</param>
+        /// <param name="cmd">The matching command</param>
+        /// <returns>True if command name matched an existing command</returns>
         internal bool Match(ReadOnlySpan<byte> command, out CustomRawStringCommand cmd)
             => rawStringCommandMap.eMapUnsafe.MatchCommand(command, out cmd);
 
+        /// <summary>
+        /// Get a custom transaction by name
+        /// </summary>
+        /// <param name="command">The transaction name to match</param>
+        /// <param name="cmd">The matching transaction</param>
+        /// <returns>True if transaction name matched an existing transaction</returns>
         internal bool Match(ReadOnlySpan<byte> command, out CustomTransaction cmd)
             => transactionProcMap.eMapUnsafe.MatchCommand(command, out cmd);
 
+        /// <summary>
+        /// Get a custom object command by name
+        /// </summary>
+        /// <param name="command">The command name to match</param>
+        /// <param name="cmd">The matching command</param>
+        /// <returns>True if command name matched an existing command</returns>
         internal bool Match(ReadOnlySpan<byte> command, out CustomObjectCommand cmd)
             => objectCommandMap.eMapUnsafe.MatchSubCommand(command, out cmd);
 
+        /// <summary>
+        /// Get a custom procedure by name
+        /// </summary>
+        /// <param name="command">The procedure name to match</param>
+        /// <param name="cmd">The matching procedure</param>
+        /// <returns>True if procedure name matched an existing procedure</returns>
         internal bool Match(ReadOnlySpan<byte> command, out CustomProcedureWrapper cmd)
             => customProcedureMap.eMapUnsafe.MatchCommand(command, out cmd);
 
+        /// <summary>
+        /// Get custom command info by name
+        /// </summary>
+        /// <param name="cmdName">The command name</param>
+        /// <param name="respCommandsInfo">The matching command info</param>
+        /// <returns>True if command info was found</returns>
         internal bool TryGetCustomCommandInfo(string cmdName, out RespCommandsInfo respCommandsInfo)
-        {
-            return this.customCommandsInfo.TryGetValue(cmdName, out respCommandsInfo);
-        }
+            => this.customCommandsInfo.TryGetValue(cmdName, out respCommandsInfo);
 
+        /// <summary>
+        /// Get custom command docs by name
+        /// </summary>
+        /// <param name="cmdName">The command name</param>
+        /// <param name="respCommandsDocs">The matching command docs</param>
+        /// <returns>True if command docs was found</returns>
         internal bool TryGetCustomCommandDocs(string cmdName, out RespCommandDocs respCommandsDocs)
-        {
-            return this.customCommandsDocs.TryGetValue(cmdName, out respCommandsDocs);
-        }
+            => this.customCommandsDocs.TryGetValue(cmdName, out respCommandsDocs);
 
+        /// <summary>
+        /// Get all custom command infos
+        /// </summary>
+        /// <returns>Map between custom command name and custom command info</returns>
+        internal IReadOnlyDictionary<string, RespCommandsInfo> GetAllCustomCommandsInfos()
+            => this.customCommandsInfo.AsReadOnly();
+
+        /// <summary>
+        /// Get all custom command docs
+        /// </summary>
+        /// <returns>Map between custom command name and custom command docs</returns>
+        internal IReadOnlyDictionary<string, RespCommandDocs> GetAllCustomCommandsDocs()
+            => this.customCommandsDocs.AsReadOnly();
+
+        /// <summary>
+        /// Get count of all custom command infos
+        /// </summary>
+        /// <returns>Count</returns>
+        internal int GetCustomCommandInfoCount() => customCommandsInfo.Count;
+
+        /// <summary>
+        /// Get RespCommand enum by command ID
+        /// </summary>
+        /// <param name="id">Command ID</param>
+        /// <returns>Matching RespCommand</returns>
         internal RespCommand GetCustomRespCommand(int id)
             => (RespCommand)(CustomRawStringCommandMinId + id);
 
+        /// <summary>
+        /// Get GarnetObjectType enum by object type ID
+        /// </summary>
+        /// <param name="id">Object type ID</param>
+        /// <returns>Matching GarnetObjectType</returns>
         internal GarnetObjectType GetCustomGarnetObjectType(int id)
             => (GarnetObjectType)(CustomObjectTypeMinId + id);