--- a/libs/server/Custom/CustomObjectCommand.cs
+++ b/libs/server/Custom/CustomObjectCommand.cs
@@ -11,16 +11,18 @@ namespace Garnet.server
         public readonly byte id;
         public readonly byte subid;
         public readonly CommandType type;
+        public readonly int arity;
         public readonly CustomObjectFactory factory;
         public readonly CustomObjectFunctions functions;
 
-        internal CustomObjectCommand(string name, byte id, byte subid, CommandType type, CustomObjectFactory factory, CustomObjectFunctions functions = null)
+        internal CustomObjectCommand(string name, byte id, byte subid, CommandType type, int arity, CustomObjectFactory factory, CustomObjectFunctions functions = null)
         {
             NameStr = name.ToUpperInvariant();
             this.Name = System.Text.Encoding.ASCII.GetBytes(NameStr);
             this.id = id;
             this.subid = subid;
             this.type = type;
+            this.arity = arity;
             this.factory = factory;
             this.functions = functions;
         }