--- a/libs/server/Custom/CustomTransaction.cs
+++ b/libs/server/Custom/CustomTransaction.cs
@@ -12,15 +12,17 @@ namespace Garnet.server
 
         public readonly string NameStr;
         public readonly byte id;
+        public readonly int arity;
         public readonly Func<CustomTransactionProcedure> proc;
 
-        internal CustomTransaction(string name, byte id, Func<CustomTransactionProcedure> proc)
+        internal CustomTransaction(string name, byte id, int arity, Func<CustomTransactionProcedure> proc)
         {
             if (name == null)
                 throw new GarnetException("CustomTransaction name is null");
             NameStr = name.ToUpperInvariant();
             this.Name = System.Text.Encoding.ASCII.GetBytes(NameStr);
             this.id = id;
+            this.arity = arity;
             this.proc = proc ?? throw new GarnetException("CustomTransactionProcedure is null");
         }
     }